Add function to record read and write values for AugAssign node
This function is added to record node id in read_lines and
write_lines. Existing system was not able to detect sum += 1.
It is able detect sum = sum + 1 but it adds sum only in the write_lines.
This function will add sum into read_lines and write_lines because
statement is reading the old value of sum and writing the new value
in the same statement.

Signed-off-by: Akash Desai <akashdesai2007@gmail.com>