Fix up for function calls on objects.
This fix up will the record the name of an object it is declared
previously before calling the function. For example, if statement
is  values.append(1) then "values" should be recorded in the
read_lines and write_lines as well  because this statement is reading
already declared value of "values" and appending data to it.

Assuming function calls will modify the object.

Signed-off-by: Akash Desai <akashdesai2007@gmail.com>